+++ Firestore/core/src/remote/datastore.cc
@@ -166,26 +166,27 @@ void Datastore::CommitMutations(const std::vector<Mutation>& mutations,
                                 CommitCallback&& callback) {
   ResumeRpcWithCredentials(
       // TODO(c++14): move into lambda.
-      [this, mutations,
-       callback](const StatusOr<AuthToken>& maybe_credentials) mutable {
-        if (!maybe_credentials.ok()) {
-          callback(maybe_credentials.status());
+      [this, mutations, callback](const StatusOr<AuthToken>& auth_token,
+                                  const std::string& app_check_token) mutable {
+        if (!auth_token.ok()) {
+          callback(auth_token.status());
           return;
         }
-        CommitMutationsWithCredentials(maybe_credentials.ValueOrDie(),
+        CommitMutationsWithCredentials(auth_token.ValueOrDie(), app_check_token,
                                        mutations, std::move(callback));
       });
 }
 
 void Datastore::CommitMutationsWithCredentials(
-    const AuthToken& token,
+    const credentials::AuthToken& auth_token,
+    const std::string& app_check_token,
     const std::vector<Mutation>& mutations,
     CommitCallback&& callback) {
   grpc::ByteBuffer message =
       MakeByteBuffer(datastore_serializer_.EncodeCommitRequest(mutations));
 
   std::unique_ptr<GrpcUnaryCall> call_owning = grpc_connection_.CreateUnaryCall(
-      kRpcNameCommit, token, std::move(message));
+      kRpcNameCommit, auth_token, app_check_token, std::move(message));
   GrpcUnaryCall* call = call_owning.get();
   active_calls_.push_back(std::move(call_owning));
 
@@ -206,27 +207,28 @@ void Datastore::LookupDocuments(const std::vector<DocumentKey>& keys,
                                 LookupCallback&& callback) {
   ResumeRpcWithCredentials(
       // TODO(c++14): move into lambda.
-      [this, keys,
-       callback](const StatusOr<AuthToken>& maybe_credentials) mutable {
-        if (!maybe_credentials.ok()) {
-          callback(maybe_credentials.status());
+      [this, keys, callback](const StatusOr<AuthToken>& auth_token,
+                             const std::string& app_check_token) mutable {
+        if (!auth_token.ok()) {
+          callback(auth_token.status());
           return;
         }
-        LookupDocumentsWithCredentials(maybe_credentials.ValueOrDie(), keys,
-                                       std::move(callback));
+        LookupDocumentsWithCredentials(auth_token.ValueOrDie(), app_check_token,
+                                       keys, std::move(callback));
       });
 }
 
 void Datastore::LookupDocumentsWithCredentials(
-    const AuthToken& token,
+    const credentials::AuthToken& auth_token,
+    const std::string& app_check_token,
     const std::vector<DocumentKey>& keys,
     LookupCallback&& callback) {
   grpc::ByteBuffer message =
       MakeByteBuffer(datastore_serializer_.EncodeLookupRequest(keys));
 
   std::unique_ptr<GrpcStreamingReader> call_owning =
-      grpc_connection_.CreateStreamingReader(kRpcNameLookup, token,
-                                             std::move(message));
+      grpc_connection_.CreateStreamingReader(
+          kRpcNameLookup, auth_token, app_check_token, std::move(message));
   GrpcStreamingReader* call = call_owning.get();
   active_calls_.push_back(std::move(call_owning));
 
@@ -255,32 +257,58 @@ void Datastore::OnLookupDocumentsResponse(
 }
 
 void Datastore::ResumeRpcWithCredentials(const OnCredentials& on_credentials) {
-  // Auth may outlive Firestore
+  // Auth/AppCheck may outlive Firestore
   std::weak_ptr<Datastore> weak_this{shared_from_this()};
+  auto credentials = std::make_shared<CallCredentials>();
+
+  auto done = [weak_this, credentials, on_credentials](
+                  const absl::optional<StatusOr<AuthToken>>& auth,
+                  const absl::optional<std::string>& app_check) {
+    auto strong_this = weak_this.lock();
+    if (!strong_this) {
+      return;
+    }
+
+    std::lock_guard<std::mutex> lock(credentials->mutex);
+    if (auth) {
+      credentials->auth = *auth;
+      credentials->auth_received = true;
+    }
+
+    if (app_check) {
+      credentials->app_check = *app_check;
+      credentials->app_check_received = true;
+    }
+
+    if (!credentials->auth_received || !credentials->app_check_received) {
+      return;
+    }
+
+    const StatusOr<AuthToken>& auth_token = credentials->auth;
+    const std::string& app_check_token = credentials->app_check;
+
+    strong_this->worker_queue_->EnqueueRelaxed(
+        [weak_this, auth_token, app_check_token, on_credentials] {
+          auto strong_this = weak_this.lock();
+          if (!strong_this) {
+            return;
+          }
+          // In case this callback is invoked after Datastore has been shut
+          // down.
+          if (strong_this->is_shut_down_) {
+            return;
+          }
+          on_credentials(auth_token, app_check_token);
+        });
+  };
 
   auth_credentials_->GetToken(
-      [weak_this, on_credentials](const StatusOr<AuthToken>& result) {
-        auto strong_this = weak_this.lock();
-        if (!strong_this) {
-          return;
-        }
+      [done](const StatusOr<AuthToken>& auth) { done(auth, absl::nullopt); });
 
-        strong_this->worker_queue_->EnqueueRelaxed(
-            [weak_this, result, on_credentials] {
-              auto strong_this = weak_this.lock();
-              if (!strong_this) {
-                return;
-              }
-              // In case Auth callback is invoked after Datastore has been shut
-              // down.
-              if (strong_this->is_shut_down_) {
-                return;
-              }
-
-              on_credentials(result);
-            });
+  app_check_credentials_->GetToken(
+      [done](const StatusOr<std::string>& app_check) {
+        done(absl::nullopt, app_check.ValueOrDie());  // AppCheck never fails
       });
-  // TODO(appcheck): Fetch AppCheck token
 }
 
 void Datastore::HandleCallStatus(const Status& status) {