+++ Firestore/core/src/remote/datastore.h
@@ -156,23 +156,33 @@ class Datastore : public std::enable_shared_from_this<Datastore> {
   }
 
  private:
+  struct CallCredentials {
+    mutable std::mutex mutex;
+    std::string app_check;
+    bool app_check_received = false;
+    util::StatusOr<credentials::AuthToken> auth;
+    bool auth_received = false;
+  };
+
   void PollGrpcQueue();
 
   void CommitMutationsWithCredentials(
-      const credentials::AuthToken& token,
+      const credentials::AuthToken& auth_token,
+      const std::string& app_check_token,
       const std::vector<model::Mutation>& mutations,
       CommitCallback&& callback);
 
   void LookupDocumentsWithCredentials(
-      const credentials::AuthToken& token,
+      const credentials::AuthToken& auth_token,
+      const std::string& app_check_token,
       const std::vector<model::DocumentKey>& keys,
       LookupCallback&& callback);
   void OnLookupDocumentsResponse(
       const util::StatusOr<std::vector<grpc::ByteBuffer>>& result,
       const LookupCallback& callback);
 
-  using OnCredentials =
-      std::function<void(const util::StatusOr<credentials::AuthToken>&)>;
+  using OnCredentials = std::function<void(
+      const util::StatusOr<credentials::AuthToken>&, const std::string&)>;
   void ResumeRpcWithCredentials(const OnCredentials& on_credentials);
 
   void HandleCallStatus(const util::Status& status);